chore(EventDetails): Extract event save button into single composable
